Clean directories
Changed the names and architecture of every files in the project to
follow standards.
