+++ src/main/java/motion/MoveStrategyCrate.java
