+++ src/test/java/overlap/SokobanOverlapRulesApplierTest.java
